Begins work on proper formatting of data
As it currently stands, the data for "ACCOUNTS.TXT" is being read into its data form incorrectly. Using Copilot, I was getting the data to have more spaces and even including the newline character at the end of each entry, thus completely messing up its proper format reading. However, using the same line of code I found that "TRANSACTIONS.TXT" had no such problems and was read in appropriately. Neither of which in the data stream are being read in as a whole list however, which might be a problem worth investigating to debug this section of code prior to data manipulation. I have swapped between Copilot and ChatGPT for different ideas, and am seeing if using both might lead to better results.