Refactors Input File Reading
This segment is far from completed, but the main function responsible for reading in my data is in desperate need of being refactored and simplified into multiple separate functions. Much of the issues I'm dealing with in regards of having the function be able to parse its information appropriately to further manipulation down the line is due to the input function being too robust in its current form, doing too many functions in one and creating my own pyramid of doom with even less composition. A .txt of the notes from Prof. Hare are on a flash drive to better help me correct the problem, but I'll need to seriously spend some time on ensuring the logic has been separated out and thorough enough to stand on its own before I can implement the logic necessary to get this project done.