Adds test tab and begins refactoring reading_accounts
Test tab is a side tab that I am using to ensure that the code and logic works before integrating it into the main work. Right now I am working on matching the data to the struct appropriately before I stick it into a master list for all accounts that exist.