Partial Updates tests.rkt file into test suites and cases
Preprocessing was edited to include a new id tag, as outlined in the original assignment. The tests.rkt file is being refactored to reflect separate test suites and cases for both readability and ease of testing. Refactoring is not necessary in this stage, but definitely helps in assisting on my maintenance of the code while I continue to finalize the product.