Last Stage of Development
This is down to the wire, with the tests.rkt file being the only part holding me back from submitting since I want to ensure it is right before submitting. Without the proper unit test to verify the output, I don't feel super comfortable with submitting even with the deadline being so close to what it is. Everything else however is already prepared and ready to go, with the unit tests verifying that my uses are appropriate and with my close eye following the outline for the project to ensure all elements have been included appropriately.